Modularize application: move app components to 'app' directory
